Update README.md
Well as we know now a days there are various OTT platforms  combining both three major platforms i.e; Amazon Prime , Netflix , disney +hotstar & the local platforms like voot etc. It sometimes becomes difficult for the users to decide which platform or what to watch.

So, to resolve this issue in the project a webpage is made which integrated with Chat bot created by the help of IBM Watson Assistant and IBM cloud . The data visualization is done by the help of IBM Cognos Analytics.